refactor: 게임 승패를 구하는 기능을 분리
- BlackjackGame에서 게임 승패를 구하는 로직을 GameResult로 분리
